Rename `cookie-universal-nuxt` plugin to `cookie-universal`
